Move the color definitions in a separate file
And document the whole lot.
